requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
Brotli==1.1.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask==3.1.0
fonttools==4.56.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.8
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
msgspec==0.19.0
numpy==2.2.4
openai==1.73.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rq
redis
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
zopfli==0.2.3.post1
